==> 1. data_extraction.ipynb <==
# KOL Data Standardization - Step 1 (Data Extraction)

"""
KOL Data Standardization - Step 1 (Extract the Data from various sources (Batches) and store it in the Landing table)

This module represents the first step in KOL Data Standardization process where we load the data which is available to us
in different set of batches. The data once loaded is checked for null values with following constraints:
- The mdm_id should not be null.
- The first name of the KOL should not be null.
Any data point failing this constraint will be stored separately (warehouse area) for future reference. Rest will be sent to landing area.
"""

import os
import datetime
import pandas as pd

RAW_DATA_STORE_PATH = "data_store"
WAREHOUSE_AREA_FILE_PATH = "output_store/landing_area.xlsx"
LANDING_AREA_FILE_PATH = "output_store/landing_area.xlsx"

# Data Extraction Process - Loading all the data from different batches and storing them in either warehouse or in landing stage.

warehouse_area = []
landing_area = []
print("Extracting the data from various batches and storing it in Landing area.\n")
for batch in os.listdir(RAW_DATA_STORE_PATH):
    file_path = os.path.join(RAW_DATA_STORE_PATH, batch)
    print(f"Processing batch: {file_path}")
    dataframe = pd.read_excel(file_path)

    condition = dataframe[["mdm_id", "first_name"]].isna().any(axis=1)

    # Datapoints having NULL values in essential columns
    warehouse_area.append(dataframe[condition])

    # Adequate data input for landing stage
    correct_data_df = dataframe[~condition].reset_index(drop=True)
    correct_data_df["batch_id"] = datetime.datetime.now()
    landing_area.append(correct_data_df)

warehouse_df = pd.concat(warehouse_area).reset_index(drop=True)
landing_df = pd.concat(landing_area).reset_index(drop=True)
print("\nSuccessfully completed data extraction process")

# Storing both the dataframes
warehouse_df.to_excel(WAREHOUSE_AREA_FILE_PATH, index=False)
landing_df.to_excel(LANDING_AREA_FILE_PATH, index=False)
print("\nSuccessfully updated the Landing area")

# Warehouse DataFrame
warehouse_df.head()

# Landing Area DataFrame
landing_df.head()
