==> 3. data_loading.ipynb <==
# KOL Data Standardization - Step 3 (Data Loading)

"""
KOL Data Standardization - Step 3 (Extracting the staging area data and loading it to the reporting area)

This module represents the final step in KOL Data Standardization process where we load the staging area data and make it adequate for the reporting layer.
It includes the following steps:
- Loading the staging area dataframe into the memory.
- Separating out the mdm_ids and their degree/speciality mappings to store in separate tables.
- Store the data into the reporting layer.
The above steps will ensure that the data is ready for the reporting layer which is suitable for the end-users.
"""

import pandas as pd
from IPython.display import display

STAGING_AREA_FILE_PATH = "output_store/staging_area.xlsx"
REPORTING_AREA_FILE_PATH = "output_store/reporting_area.xlsx"
REPORTING_AREA_SP_MAP_FILE_PATH = "output_store/reporting_area_sp_map.xlsx"
REPORTING_AREA_DEG_MAP_FILE_PATH = "output_store/reporting_area_deg_map.xlsx"

print("TASK: Loading the Staging Area Dataframe")
staging_df = pd.read_excel(STAGING_AREA_FILE_PATH)
display(staging_df.head())

print("TASK: Preparing the KOL speciality and degree maps")
kol_speciality_map = staging_df.assign(speciality_id=staging_df['speciality'].str.split(',')).explode('speciality_id')[["mdm_id", "speciality_id"]]
kol_degree_map = staging_df.assign(degree_id=staging_df['degree'].str.split(',')).explode('degree_id')[["mdm_id", "degree_id"]]
print("TASK: Preparing the KOL reporting layer dataset")
reporting_df = staging_df.drop(["speciality", "degree", "batch_id"], axis=1)

display(kol_speciality_map.head())

display(kol_degree_map.head())

display(reporting_df.head())

print("TASK: Saving the data maps and reporting area dataset")
kol_speciality_map.to_excel(REPORTING_AREA_SP_MAP_FILE_PATH, index=False)
kol_degree_map.to_excel(REPORTING_AREA_DEG_MAP_FILE_PATH, index=False)
reporting_df.to_excel(REPORTING_AREA_FILE_PATH, index=False)
print("TASK: Successfully loaded the data to the reporting area")
